Remove unnecessary if-statement
Function explode always returns an array. The case that this array
contains a single empty string is already caught in the loop. Even if
the array ever was empty, using a foreach loop on an empty array results
in zero executions of the loop body.

So there is no need to have the extra if-statement around the foreach
loop.