Allow regular expressions in excludes
The configuration settings have been talking about regular expressions
in internal/external URI patterns for a long time, but the patterns used
to be substrings. They could not even appear at the beginning of the
URI. Assume that the latter was a mistake. (This is relevant for the
upgrade procedure.)

Change the configuration settings to use real regular expressions, and
improve their explanatory texts.

Add an upgrade procedure which converts the static strings used up to
now to regular expressions.